+++ modules/juce_core/zip/zlib/trees.c
@@ -169,7 +169,7 @@ local void gen_trees_header OF((void));
 
 #else /* DEBUG */
 #  define send_code(s, c, tree) \
-     { if (z_verbose>2) fprintf(stderr,"\ncd %3d ",(c)); \
+     { if (z_verbose>2) fprintf(stderr,"\ncd %3d ",(int)(c)); \
        send_bits(s, tree[c].Code, tree[c].Len); }
 #endif
